C12: Fix connectivity issue with config-server
Config client first loads default profile and try to connect to
default profile's uri. And since default profile tries to connect
localhost and localhost can not be reached from docker containers,
it fails, resulting in application shutdown.

So, now we provide the uri through envrionment variable with
localhost as default so that it can work when running in non
dockerized envrionment.